Add character killed button
Character killed button will simulate the character's death and delete
the character sheet in the future. Button added now for styling and to
add future functionality.
Replace the idomatic text element with a span element to hold the
skull icon inside the button. It's a more appropriate element for this
use as the purpose is to apply the icon's class styling.